refactor(optimize): give the value-copy peel an infrastructure home
`strip_one_value_copy` sees through a `$value_copy$T` wrapper — an arena
query, not a container-SROA concept. Living in `container_sroa` made
`sroa` reach across to a sibling pass for it, the only such dependency in
the SROA family. Moved next to `strip_refs` in `arena_query`, where the
other see-through-a-wrapper query already lives; both passes now depend
on the shared query instead of on each other.
